requirements.txt
mutagen==1.47.0
pillow==10.1.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
jinja2==3.1.2

